Introduce UnitContext class
